moved enums to enums folder
