+++ source/coordinator/GreenEnergyHub.Aggregation.Application/Coordinator/InputProcessor.cs
@@ -20,6 +20,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using GreenEnergyHub.Aggregation.Application.Coordinator.Interfaces;
 using GreenEnergyHub.Aggregation.Domain.DTOs.MetaData;
+using GreenEnergyHub.Aggregation.Domain.DTOs.MetaData.Enums;
+
 using Microsoft.Extensions.Logging;
 using NodaTime;
