+++ source/coordinator/GreenEnergyHub.Aggregation.Application/Coordinator/TriggerBaseArguments.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections.Generic;
 using GreenEnergyHub.Aggregation.Application.Utilities;
-using GreenEnergyHub.Aggregation.Domain.DTOs.MetaData;
+using GreenEnergyHub.Aggregation.Domain.DTOs.MetaData.Enums;
 using NodaTime;
 
 namespace GreenEnergyHub.Aggregation.Application.Coordinator
